Cardinal port: put all Venom code in a namespace
Because all the various plugins are static-linked
together in Cardinal, it is much safer to put symbols
in a namespace instead of populating the global namespace.

So now, all code is wrapped in namespace Venom { }.

Only the "modelVenom..." global variables are still in
the global namespace, but the "Venom" in them serves
as a de facto namespace.